Split monolithic gig endpoints into multiple focused endpoint files
